moved the loading of existing-payment-options into contribtionsLandingInit.js (rather than it happening for every page) - in future when rolled out to further checkouts we might want to move to a new sharedLandingInit.js or something like that
